game has working music
music was implemented as well a a game manager that keeps track of the music that needs to be played, the arrow holder and a bool to track that the scene can start. The beat scroller was change to have that bool to start moving to be change from the game manger rather then the script itself looking for the input. Music being used is city Funk from tape girl mod for FNF so this song is not mine but its for learning perposes so hopefully will be changed at a later time