temporary missed and hit implemented
the game manager now has functions that run when a note is hit or missed. It currently only prints text to the console log to see if they react when they are supposed to. the arrow behavior then calls the static instance of game manager to use the functions created for the notes hit and missed to be called when appropriate.